sweep: move tx generation to separate file
